pyproject.toml
[project]
name = "use_your_head"
version = "0.1.0"
description = "Package for solving Tetris-like puzzles"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anywidget>=0.9.13",
    "fire>=0.7.0",
    "ipymario>=0.1.0",
    "marimo>=0.11.0",
    "pydantic>=2.10.5",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = ["pytest>=8.3.4"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
dlx = "use_your_head.__main__:main"
